+++ lib/recurly/plan.php
@@ -19,7 +19,7 @@ class Recurly_Plan extends Recurly_Resource
       'bypass_hosted_confirmation','unit_name','payment_page_tos_link',
       'plan_interval_length','plan_interval_unit','trial_interval_length',
       'trial_interval_unit','unit_amount_in_cents','setup_fee_in_cents',
-      'total_billing_cycles','accounting_code'
+      'total_billing_cycles','accounting_code','tax_exempt'
     );
     Recurly_Plan::$_nestedAttributes = array(
       'add_ons'